Make sure search bundle.total is a number (#479)

+++ packages/server/src/fhir/repo.ts
@@ -496,7 +496,7 @@ export class Repository {
   async #getTotalCount(searchRequest: SearchRequest): Promise<number> {
     const client = getClient();
     const builder = new SelectQuery(searchRequest.resourceType).raw(
-      `COUNT (DISTINCT "${searchRequest.resourceType}"."id") AS "count"`
+      `COUNT (DISTINCT "${searchRequest.resourceType}"."id")::int AS "count"`
     );
 
     this.#addDeletedFilter(builder);
